==> DataBuildingScripts/4-dim_red.ipynb <==
import pandas as pd
from sklearn.decomposition import PCA

crime_num = pd.read_csv("../Data/FE/crime_num_fe.csv")
crime_violent = pd.read_csv("../Data/FE/crime_violent_fe.csv")


# PCA
I want to group the variables that might be most related to each other, such as temp and precip vars. 
Each group become one column, and when doing modeling, will be used instead of the features it represents. 
This will be useful to determine if model prectibability improves with less correlation. 

crime_num.columns

X_temp = crime_num[['tempmax', 'tempmin', 'temp', 'feelslikemax', 'feelslikemin',
       'feelslike', 'dew']]

X_precip = crime_num[['precip', 'precipprob', 'precipcover', 'snow', 'snowdepth']]

X_wind = crime_num[['windgust', 'windspeed', 'winddir']]

X_cloud = crime_num[['visibility', 'cloudcover', 'humidity']]

X_radiation = crime_num[['solarradiation', 'solarenergy', 'uvindex', 'severerisk']]

pca = PCA(n_components=1)
X_pca_temp = pca.fit(X_temp)
X_pca_temp = X_pca_temp.fit_transform(X_temp)
X_pca_precip = pca.fit(X_precip)
X_pca_precip = X_pca_precip.fit_transform(X_precip)
X_pca_wind = pca.fit(X_wind)
X_pca_wind = X_pca_wind.fit_transform(X_wind)
X_pca_cloud = pca.fit(X_cloud)
X_pca_cloud = X_pca_cloud.fit_transform(X_cloud)
X_pca_radiation = pca.fit(X_radiation)
X_pca_radiation = X_pca_radiation.fit_transform(X_radiation)


crime_num['temp_pca'] = X_pca_temp
crime_num['precip_pca'] = X_pca_precip
crime_num['wind_pca'] = X_pca_wind
crime_num['temp_pca'] = X_pca_cloud
crime_num['radiation_pca'] = X_pca_radiation


crime_violent['temp_pca'] = X_pca_temp
crime_violent['precip_pca'] = X_pca_precip
crime_violent['wind_pca'] = X_pca_wind
crime_violent['cloud _pca'] = X_pca_cloud 
crime_violent['radiation_pca'] = X_pca_radiation


crime_num = crime_num.drop(columns=X_temp.columns.tolist() + X_precip.columns.tolist() + X_wind.columns.tolist() + X_cloud.columns.tolist() + X_radiation.columns.tolist())
crime_violent = crime_violent.drop(columns=X_temp.columns.tolist() + X_precip.columns.tolist() + X_wind.columns.tolist() + X_cloud.columns.tolist() + X_radiation.columns.tolist())


crime_num.to_csv("../Data/FE/crime_num_dr.csv")
crime_violent.to_csv("../Data/FE/crime_violent_dr.csv")
