==> DataBuildingScripts/5-lasso_selection.ipynb <==
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.linear_model import Lasso

import pandas as pd
import numpy as np

crime_num = pd.read_csv("../Data/FE/crime_num_fe.csv")
crime_violent = pd.read_csv("../Data/FE/crime_violent_fe.csv")

crime_num.columns

y = crime_num['Num_Crimes']
X = crime_num.drop(columns=['Num_Crimes','TS_interval','Date','Day','Month', 'Year','tempmax', 'tempmin', 'feelslikemax', 'feelslikemin', 'dew'])
X['Holiday'] = X['Holiday'] != "None"

# Lasso Variable Selection
Unlike with PCA/dimension reduction, this aim is to determine if filtering the data to certain columns improve predictions

categorical_cols = X.select_dtypes(include=['object', 'category']).columns.tolist()
numeric_cols = X.select_dtypes(include=['int64', 'float64']).columns.tolist()

# Create a transformer for numeric data
numeric_transformer = Pipeline(steps=[
    ('scaler', StandardScaler())
])

# Create a transformer for categorical data
categorical_transformer = Pipeline(steps=[
    ('onehot', OneHotEncoder(handle_unknown='ignore'))
])

# Combine transformers into a preprocessor
preprocessor = ColumnTransformer(
    transformers=[
        ('num', numeric_transformer, numeric_cols),
        ('cat', categorical_transformer, categorical_cols)
    ])

lasso_model = Pipeline(steps=[
    ('preprocessor', preprocessor),
    ('regressor', Lasso(alpha=0.05, random_state=0)) # .05 is subjective, not a model I am predicting so I didn't worry about it at this stage.
])

# Train the Lasso model
lasso_model.fit(X, y)

# Extract Feature Names

feature_names = (lasso_model.named_steps['preprocessor']
                           .transformers_[1][1]
                           .named_steps['onehot']
                           .get_feature_names_out())

# Extend this list with the names of numeric columns
feature_names = np.concatenate([numeric_cols, feature_names])

# Get coefficients from the lasso model
coefficients = lasso_model.named_steps['regressor'].coef_

# Print nonzero coefficients and their corresponding feature names
print("Non-zero coefficients and their corresponding features:")
for coef, feat in zip(coefficients, feature_names):
    print(f"{feat}: {coef}")

nonzero_coefs = []
for coef, feat in zip(coefficients, feature_names):
    if coef != 0:
        nonzero_coefs.append(feat)

print(nonzero_coefs)
len(nonzero_coefs)

nonzero_coefs = ['feelslike', 'humidity', 'precip', 'precipcover', 'snow', 'snowdepth', 'sealevelpressure', 'cloudcover', 'solarradiation', 'solarenergy', 'severerisk', 'suntime', 'Time_Interval']

filt_num = crime_num[['Num_Crimes', 'TS_interval', 'Date', 'Day', 'Month', 'Year'] + nonzero_coefs]
filt_violent = crime_violent[['Num_Crimes', 'TS_interval', 'Date', 'Day', 'Month', 'Year'] + nonzero_coefs]

filt_num.to_csv("../Data/FE/crime_num_varselect.csv")
filt_violent.to_csv("../Data/FE/crime_violent_varselect.csv")
